+++ src/main/java/net/geforcemods/securitycraft/compat/jei/SecurityCraftToVanillaCategory.java
@@ -2,12 +2,16 @@ package net.geforcemods.securitycraft.compat.jei;
 
 import mezz.jei.api.constants.VanillaTypes;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
+import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
 import mezz.jei.api.helpers.IGuiHelper;
 import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeIngredientRole;
 import mezz.jei.api.recipe.types.IRecipeType;
 import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.GuiGraphicsExtractor;
 import net.minecraft.network.chat.Component;
+import net.minecraft.util.CommonColors;
 import net.minecraft.world.item.ItemStack;
 
 public class SecurityCraftToVanillaCategory extends BaseCategory {
@@ -17,12 +21,11 @@ public class SecurityCraftToVanillaCategory extends BaseCategory {
 		super(helper);
 	}
 
-// TODO: Broken in JEI
-//	@Override
-//	public void draw(ReinforcerRecipe recipe, IRecipeSlotsView recipeSlotsView, GuiGraphics guiGraphics, double mouseX, double mouseY) {
-//		super.draw(recipe, recipeSlotsView, guiGraphics, mouseX, mouseY);
-//		guiGraphics.text(Minecraft.getInstance().font, OUTPUT_TEXT, 24, 30, CommonColors.DARK_GRAY, false);
-//	}
+	@Override
+	public void draw(ReinforcerRecipe recipe, IRecipeSlotsView recipeSlotsView, GuiGraphicsExtractor guiGraphics, double mouseX, double mouseY) {
+		super.draw(recipe, recipeSlotsView, guiGraphics, mouseX, mouseY);
+		guiGraphics.text(Minecraft.getInstance().font, OUTPUT_TEXT, 24, 30, CommonColors.DARK_GRAY, false);
+	}
 
 	@Override
 	public Component getTitle() {