+++ Changelog.md
@@ -28,3 +28,9 @@
 - Fix: The experience level and the background of the locator bar are visible while mounted to a Security Camera 
 
 --------------------------Changelog for v1.10.1-beta2 of SecurityCraft--------------------------
+
+- Fix: Screens that contain item slots do not darken the world in the background
+- Fix: SecurityCraft's custom JEI recipe categories do not show up correctly
+- Fix: The Jade and WTHIT overlays show while viewing a Security Camera
+- Fix: Water in a Reinforced Cauldron is tinted in reinforced blocks tint
+- Misc.: The minimum required NeoForge version is 26.1.1.8-beta